[1428/1512] Migrate to numpy v2 and zarr v3 with updated test_data path format [all tests ci] (#1531)
* [1428/1512] Migrate to numpy v2 and zarr v3 with updated test_data path formatting

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* [1428/1512] Fix issues with migration to numpy v2 and zarr v3

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* [1428/1512] Fix zarr store delete and create and resolve some migration warnings

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* [1428/1512] Reinstall netcdf in github PR workflow to resolve numpy.dtype size changed error

* [1428-1512] Remove FSStore type from zarr store delete

* [1541] Remove python 3.10 support

* force zarr>=3

* [1531] Minor changes to support zarr v3

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* [1531] Adjust to follow line length limit

* [1531] Test removing github workflow netcdf workaround

---------

+++ echopype/convert/parse_ad2cp.py
@@ -319,7 +319,7 @@ class Ad2cpDataPacket:
         Calculates and returns the timestamp of the packet
         """
 
-        year = self.data["year"] + 1900
+        year = self.data["year"].astype(np.uint16) + 1900
         month = self.data["month"] + 1
         day = self.data["day"]
         hour = self.data["hour"]
